delta time is added into the game how a mechanism to represent real time in game

+++ src/Game/Gameplay/Events/WeaponEvents.java
@@ -55,7 +55,7 @@ public final class WeaponEvents {
      * @param weapon          el arma que recarga
      * @param reloadTimeTicks duración total de la recarga en ticks
      */
-    public record OnReloadStart(ModifiedWeapon weapon, int reloadTimeTicks) {}
+    public record OnReloadStart(ModifiedWeapon weapon, double reloadTimeTicks) {}
 
     /**
      * Emitido cuando la recarga se completa y el cargador está lleno.
